Agents/simple-calculator-agent/requirements.txt
# AgentCore deployment
bedrock-agentcore

# Strands SDK
strands-agents
strands-agents-tools

# AWS SDK
boto3
